inegración de cmd en bd
