+++ cli/commands/cmd_db.py
@@ -4,7 +4,11 @@ CLI Base de Datos
 
 import os
 import sys
+import csv
+import re
 
+from pathlib import Path
+from datetime import datetime
 import click
 from dotenv import load_dotenv
 
@@ -44,6 +48,7 @@ database.app = app
 load_dotenv()  # Take environment variables from .env
 
 entorno_implementacion = os.environ.get("DEPLOYMENT_ENVIRONMENT", "develop").upper()
+DIR_RESPALDO_BD = os.getenv("DIR_RESPALDO_BD")
 
 
 @click.group()
@@ -122,8 +127,76 @@ def copiar():
     click.echo("Termina copiar.")
 
 
+@click.command()
+@click.argument("anio_mes", type=str)
+@click.option("--output", default="reporte_respaldos_BD.csv", type=str, help="Archivo CSV a escribir")
+def generar_csv(anio_mes, output):
+    """Reporte de respaldos de BD a un archivo CSV"""
+    # Validar el archivo CSV a escribir, que no exista
+    ruta = Path(output)
+    if ruta.exists():
+        click.echo(f"AVISO: {output} existe, no voy a sobreescribirlo.")
+        return
+    # Validar que el parámetro mes_int sea YYYY-MM
+    if not re.match(r"^\d{4}-\d{2}$", anio_mes):
+        click.echo(f"ERROR: {anio_mes} no es una fecha valida (YYYY-MM)")
+        return
+    # Consultar
+    click.echo("Elaborando reporte de respaldos BD...")
+    # Lectura de archivos dentro del directorio de respaldo
+    archivos = []
+    nombre_buscado = f"pjecz_plataforma_web-{anio_mes[0:4]}-{anio_mes[5:7]}-"
+    for archivo in Path(DIR_RESPALDO_BD).rglob("*"):
+        # Consultar archivos que tengan el nombre tal
+        if nombre_buscado in archivo.name:
+            stat = os.stat(archivo)
+            patron_fecha = re.search(r"\d{4}-\d{2}-\d{2}-\d{4}", archivo.name)
+            patron_fecha = patron_fecha.group()
+            fecha = datetime.strptime(patron_fecha, "%Y-%m-%d-%H%M")
+            archivo = {
+                "fecha": fecha,
+                "nombre": archivo.name,
+                "tamanio": stat.st_size / (1024 * 1024),
+            }
+            archivos.append(archivo)
+
+    # si no hay contenido salir
+    if len(archivos) <= 0:
+        click.echo("  No se genero archivo de reporte porque no hubo registros.")
+        return
+
+    # Escribir el reporte en el archivo CSV
+    with open(ruta, "w", encoding="utf8") as puntero:
+        reporte = csv.writer(puntero)
+        reporte.writerow(
+            [
+                "Fecha",
+                "Nombre del archivo",
+                "Tamanio",
+            ]
+        )
+        # Escribir contenido
+        for archivo in archivos:
+            # Escribir la linea
+            reporte.writerow(
+                [
+                    archivo["fecha"].strftime("%Y-%m-%d %H:%M:%S"),
+                    archivo["nombre"],
+                    f"{archivo['tamanio']:.2f}",
+                ]
+            )
+
+    # Mostrar en pantalla resultado
+    click.echo(f"  |      Fecha       | Nombre                                      | Tamanio (MB)")
+    for archivo in archivos:
+        click.echo(f"  + {archivo['fecha'].strftime('%Y-%m-%d %H:%M')} | {archivo['nombre']} | {archivo['tamanio']:.2f}")
+
+    click.echo(f"  {len(archivos)} archivos de respaldo en {ruta.name}")
+
+
 cli.add_command(inicializar)
 cli.add_command(alimentar)
 cli.add_command(reiniciar)
 cli.add_command(respaldar)
 cli.add_command(copiar)
+cli.add_command(generar_csv)
